Remove old experiment runner
