--- app/services/experimentation/runner_v2.rb
@@ -1,12 +1,0 @@
-module Experimentation
-  class RunnerV2
-    def self.call
-      unless Flipper.enabled?(:experiment)
-        Rails.logger.info("Experiment feature flag is off. Experiments #{name} will not be started.")
-        return
-      end
-
-      [CurrentPatientExperiment, StalePatientExperiment].each { |experiment| experiment.conduct_daily(Date.current) }
-    end
-  end
-end
